base/serializers.py
from rest_framework import serializers
from .models import Category, Product, Order, OrderDetail

class CategorySerializer(serializers.ModelSerializer):
  class Meta:
    model = Category
    fields = '__all__'

class ProductSerializer(serializers.ModelSerializer):
  class Meta:
    model = Product
    fields = '__all__'

class OrderSerializer(serializers.ModelSerializer):
  class Meta:
    model = Order
    fields = '__all__'

class OrderDetailSerializer(serializers.ModelSerializer):
  class Meta:
    model = OrderDetail
    fields = '__all__'